requirements.txt
# Market Radar - Python Dependencies
# ==================================

# Core dependencies
feedparser>=6.0.10        # RSS feed parsing
requests>=2.31.0          # HTTP requests
python-dotenv>=1.0.0      # .env file support
pydantic>=2.0.0           # Data validation

# Market data providers
yfinance>=0.2.28          # Yahoo Finance (fallback, no API key needed)
# finnhub-python>=2.4.18  # Finnhub (optional, uncomment if using)
# polygon-api-client      # Polygon (optional, uncomment if using)

# Telegram notifications
# No extra package needed - uses requests

# Data analysis (optional, for ticker filter)
pandas>=2.0.0             # For downloading S&P 500 list
# lxml>=4.9.0             # For pandas HTML parsing (optional)

# Async/performance (optional)
# tenacity>=8.2.3         # Retry logic (already included in notifier)

# Database (built-in)
# sqlite3 is built into Python

# Testing (development only)
# pytest>=7.4.0
